+++ apps/web/src/features/record/hooks/useWebViewBirdgeListener.ts
@@ -1,0 +1,29 @@
+// hooks/useWebViewBridgeListener.ts
+import { useEffect } from 'react'
+
+type MessageHandler = (payload: string) => void
+
+export const useWebViewBridgeListener = (onImageReceived: MessageHandler) => {
+  useEffect(() => {
+    const listener = (event: MessageEvent) => {
+      const raw = typeof event.data === 'string' ? event.data : '{}'
+      const parsed = JSON.parse(raw) as { type?: string; payload?: string }
+
+      if (parsed.type === 'image' && typeof parsed.payload === 'string') {
+        onImageReceived(parsed.payload)
+
+        // 디버깅용 echo
+        window.ReactNativeWebView?.postMessage(
+          JSON.stringify({
+            from: 'web',
+            type: 'echo_log',
+            message: parsed.payload.slice(0, 50),
+          }),
+        )
+      }
+    }
+
+    window.addEventListener('message', listener)
+    return () => window.removeEventListener('message', listener)
+  }, [onImageReceived])
+}
